+++ de.tu_bs.cs.isf.e4cf.featuremodel.synthesis/src/de/tu_bs/cs/isf/e4cf/featuremodel/synthesis/widgets/WordCounter.java
@@ -2,7 +2,7 @@ package de.tu_bs.cs.isf.e4cf.featuremodel.synthesis.widgets;
 
 public class WordCounter {
 	public String word;
-	public int count = 1;
+	public double count = 1.0f;
 
 	public WordCounter(String word) {
 		this.word = word;
